+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/OnboardingHeader/index.tsx
