+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/OnboardingPageContent.styled.ts
