+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/OnboardingProgress/OnboardingProgress.styled.ts
