+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/OnboardingProgress/index.tsx
