+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/StepDescription/index.tsx
