+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/StepGuide/index.tsx
