+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/Steps/Step1/Step1.styled.ts
