+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/Steps/Step1/index.tsx
