+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/Steps/Step2/Step2.styled.ts
