+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/Steps/Step2/index.tsx
