+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/Steps/Step3/Step3.styled.ts
