+++ frontend/src/pages/OnboardingPage/components/OnboardingPageContent/Steps/Step3/index.tsx
